requirements.txt
Django==3.2.15
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.13.1
attrs==22.1.0
django-rest-knox==1.0.0
certifi==2022.6.15
cffi==1.15.1
Pillow==9.2.0
requests==2.28.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.11
wsproto==1.1.0
MarkupSafe==2.1.1
Pillow==9.2.0
pycparser==2.21
pyhcl==0.4.4
pyOpenSSL==23.0.0
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.1
s3transfer==0.6.0
shortuuid==1.0.11
typing_extensions==4.4.0
uritemplate==4.1.1
validators==0.20.0
zipp==3.11.0
